luidTableGrid [enh] Clean Code
